+++ chainladder/__init__.py
@@ -6,10 +6,12 @@ Created on Sat Jun 17 07:08:08 2017
 """
 
 
-from chainladder.UtilityFunctions import load_dataset, parallelogram_OLF
+from chainladder.utility_functions import *
 #from chainladder.stochastic.Chainladder import Chainladder, WRTO
 from chainladder.triangle import Triangle, Exposure
-from chainladder.stochastic.MackChainladder import MackChainladder
-from chainladder.stochastic.MunichChainladder import MunichChainladder
-from chainladder.stochastic.BootstrapReserve import BootChainladder
-from chainladder.deterministic import *
+#from chainladder.stochastic.MackChainladder import MackChainladder
+#from chainladder.stochastic.MunichChainladder import MunichChainladder
+#from chainladder.stochastic.BootstrapReserve import BootChainladder
+from chainladder.development import *
+from chainladder.tails import *
+from chainladder.methods import *
