--- chainladder/deterministic/__init__.py
@@ -1,12 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Created on Sat Jun 17 07:08:08 2017
-
-@author: jboga
-"""
-
-
-
-from chainladder.deterministic.chainladder import Chainladder
-from chainladder.deterministic.development import Development
-from chainladder.deterministic.benktander import Benktander, BornFerg
