+++ chainladder/development/munich.py
@@ -1,0 +1,5 @@
+''' Munich Chainladder
+Use constant.py as a template
+reference:
+    https://www.casact.org/pubs/forum/13fforum/02-Tail-Factors-Working-Party.pdf
+'''
