+++ chainladder/methods/__init__.py
@@ -1,0 +1,11 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Sat Jun 17 07:08:08 2017
+
+@author: jboga
+"""
+
+
+
+from chainladder.methods.chainladder import Chainladder
+from chainladder.methods.benktander import Benktander, BornFerg
