+++ chainladder/tails/__init__.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Sat Jun 17 07:08:08 2017
+
+@author: jboga
+"""
+
+from chainladder.tails.base import TailBase
+from chainladder.tails.constant import Constant
+from chainladder.tails.exponential import Exponential
